requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.2.0
python-decouple==3.1
pytz==2017.3
static3==0.7.0
django-etc==1.3.0
django-filter==2.4.0
django-ckeditor==6.1.0
